Refactor: 이미지 경로 상수 따로 관리
