+++ frontend/src/components/modals/DashboardBasicSettingModal/index.tsx
@@ -9,7 +9,8 @@ import InputCommon from 'src/components/inputs/Common';
 import ProfileImage from 'src/components/images/ProfileImage';
 import { Row, Col } from 'src/components/Grid';
 
-import { DEFAULT_PROFILE_IMAGE, USER_SETTING_PROFILE_IMAGE_SIZE } from 'src/globals/constants';
+import { USER_SETTING_PROFILE_IMAGE_SIZE } from 'src/globals/constants';
+import { DEFAULT_PROFILE_IMAGE } from 'src/globals/images';
 
 import userAtom from 'src/recoil/user';
 import dashboardUserInfoAtom from 'src/recoil/dashboardUserInfo';
