+++ frontend/src/components/sets/ProfileSet/index.tsx
@@ -4,10 +4,10 @@ import ProfileImageButton from 'src/components/buttons/ProfileImageButton';
 import UsernameButton from 'src/components/buttons/UsernameButton';
 
 import {
-  DEFAULT_PROFILE_IMAGE,
   POST_PROFILE_IMAGE_BUTTON_MARGIN_RIGHT,
   POST_USERNAME_BUTTON_MARGIN_RIGHT,
 } from 'src/globals/constants';
+import { DEFAULT_PROFILE_IMAGE } from 'src/globals/images';
 
 import { Wrapper } from './style';
