+++ frontend/src/globals/constants.ts
@@ -1,7 +1,3 @@
-export const DEFAULT_PROFILE_IMAGE = '/images/default_profile_image.jpg';
-export const LOGO_IMAGE = '/images/logo_image.svg';
-export const FAVICON = '/favicon.ico';
-
 // IMAGE
 export const DEFAULT_POST_IMAGE_WIDTH = 600;
 export const DEFAULT_POST_IMAGE_HEIGHT = 360;
