+++ frontend/src/globals/images.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_PROFILE_IMAGE = '/images.ts/default_profile_image.jpg';
+export const LOGO_IMAGE = '/images.ts/logo_image.svg';
+export const FAVICON = '/favicon.ico';
